Removed Try catch added "addressbook" to URL
I removed the try catch and had it just throw the exceptions "naturally". this way the code/Test will fail if the element "View Details" is not found

Also added /addressbook to the URL because otherwise it just leads me to the Xampp homepage.
Yet to work on the screenshot

Co-Authored-By: Riley Huston <98052534+mr-rjh3@users.noreply.github.com>